Rename observe to record
