+++ alphatrion/__init__.py
@@ -1,5 +1,5 @@
 from alphatrion.experiment.craft_exp import CraftExperiment
-from alphatrion.observe.observe import log_artifact, log_params
+from alphatrion.record.record import log_artifact, log_params
 from alphatrion.runtime.runtime import init
 
 __all__ = ["log_artifact", "log_params", "CraftExperiment", "init"]
